Update class-list-table.php
With PHP notices capped at 1K characters on my systems, tracking down offending module can be difficult. Having the search at line number was turning out to be a bit of hit or miss as the PHP line number didn't match the Snippet. To help narrow the list of Snippets down, I updated the line_no search to select Snippets with at least "$line_no" lines. This helped a lot. In the end, I was able to more quickly eliminate Snippets as culprits and focus on templates and plugins.